// src/enums/BetType.ts
export enum BetType {
  ANY_CRAPS = "AnyCraps",
  ANY_SEVEN = "AnySeven",
  BUY = "Buy",
  C_AND_E = "CE",
  COME = "Come",
  CRAPS_3_WAY = "Craps3Way",
  DONT_COME = "DontCome",
  DONT_PASS = "DontPass",
  FIELD = "Field",
  HARDWAY = "Hardway",
  HOP = "Hop",
  HORN = "Horn",
  HORN_HIGH = "HornHigh",
  LAY = "Lay",
  PASS_LINE = "PassLine",
  PLACE = "Place",
  PUT = "Put",
  WORLD = "World",
}